+++ src/main/java/com/qiao/picturedepot/controller/PictureController.java
@@ -4,12 +4,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.qiao.picturedepot.pojo.domain.PictureRef;
 import com.qiao.picturedepot.service.PictureService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -20,7 +22,7 @@ public class PictureController {
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     @GetMapping("/pictureRef-groups/{pictureGroupId}/pictures/{pictureId}")
-    public void getPictureFile(@PathVariable BigInteger pictureGroupId, @PathVariable BigInteger pictureId, HttpServletResponse response) throws IOException {
+    public void getPictureFile(@PathVariable Long pictureGroupId, @PathVariable Long pictureId, HttpServletResponse response) throws IOException {
         response.setHeader("Content-Type", "image/png");
 
         try(OutputStream outputStream = response.getOutputStream()){
@@ -29,7 +31,7 @@ public class PictureController {
     }
 
     @GetMapping("/pictureRef-groups/{pictureGroupId}/pictures")
-    public List<PictureRef> getPicturesOfGroup(@PathVariable BigInteger pictureGroupId){
+    public List<PictureRef> getPicturesOfGroup(@PathVariable Long pictureGroupId){
         return pictureService.getPictureListByGroup(pictureGroupId);
     }
 }
