+++ src/main/java/com/qiao/picturedepot/controller/PictureGroupController.java
@@ -6,15 +6,14 @@ import com.github.pagehelper.PageInfo;
 import com.qiao.picturedepot.dao.PictureGroupMapper;
 import com.qiao.picturedepot.exception.ServiceException;
 import com.qiao.picturedepot.pojo.domain.PictureGroup;
-import com.qiao.picturedepot.pojo.dto.PictureGroupUpdateRequest;
 import com.qiao.picturedepot.pojo.dto.PictureGroupPreviewDto;
+import com.qiao.picturedepot.pojo.dto.PictureGroupUpdateRequest;
 import com.qiao.picturedepot.service.PictureService;
 import com.qiao.picturedepot.util.ObjectUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -26,19 +25,19 @@ public class PictureGroupController {
     private PictureGroupMapper pictureGroupMapper;
 
     @GetMapping("/albums/{albumId}/pictureRef-groups")
-    public PageInfo<PictureGroupPreviewDto> getPictureGroups(@PathVariable BigInteger albumId, @RequestParam("pageNo") Integer pageNo, @RequestParam("pageSize") Integer pageSize){
+    public PageInfo<PictureGroupPreviewDto> getPictureGroups(@PathVariable Long albumId, @RequestParam("pageNo") Integer pageNo, @RequestParam("pageSize") Integer pageSize){
         PageHelper.startPage(pageNo, pageSize);
         List<PictureGroupPreviewDto> pictureGroupPreviewDtos = pictureService.getPictureGroupsByAlbum(albumId);
         return new PageInfo<>(pictureGroupPreviewDtos);
     }
 
     @GetMapping("/pictureRef-groups/{pictureGroupId}")
-    public PictureGroup getPictureGroup(@PathVariable BigInteger pictureGroupId){
+    public PictureGroup getPictureGroup(@PathVariable Long pictureGroupId){
         return pictureService.getPictureGroupById(pictureGroupId);
     }
 
     @DeleteMapping("/pictureRef-groups/{pictureGroupId}")
-    public void deletePictureGroup(@PathVariable BigInteger pictureGroupId){
+    public void deletePictureGroup(@PathVariable Long pictureGroupId){
         pictureService.deletePictureGroup(pictureGroupId);
     }
