+++ src/main/java/com/qiao/picturedepot/pojo/domain/FriendShip.java
@@ -3,13 +3,11 @@ package com.qiao.picturedepot.pojo.domain;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import java.math.BigInteger;
-
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class FriendShip extends BaseEntity{
-    private BigInteger friendUserId;
-    private BigInteger friendGroupId;
+    private Long friendUserId;
+    private Long friendGroupId;
 
     //------------------------------
     private String friendUsername;
