+++ src/main/java/com/qiao/picturedepot/pojo/dto/AcceptNewFriendRequest.java
@@ -3,11 +3,9 @@ package com.qiao.picturedepot.pojo.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Setter
 @Getter
 public class AcceptNewFriendRequest {
-    private BigInteger newFriendSystemMessageId;
+    private Long newFriendSystemMessageId;
     private String friendGroupName;
 }
