+++ src/main/java/com/qiao/picturedepot/pojo/dto/PictureGroupUpdateRequest.java
@@ -2,7 +2,6 @@ package com.qiao.picturedepot.pojo.dto;
 
 import lombok.*;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Setter
@@ -11,9 +10,9 @@ import java.util.List;
 @AllArgsConstructor
 @ToString
 public class PictureGroupUpdateRequest {
-    private BigInteger pictureGroupId;
+    private Long pictureGroupId;
     private String title;
-    private BigInteger albumId;
-    private List<BigInteger> idSequence;
-    private List<BigInteger> picturesToDelete;
+    private Long albumId;
+    private List<Long> idSequence;
+    private List<Long> picturesToDelete;
 }
