+++ src/main/java/com/qiao/picturedepot/pojo/dto/UpdateFriendInfoRequest.java
@@ -3,11 +3,9 @@ package com.qiao.picturedepot.pojo.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Setter
 @Getter
 public class UpdateFriendInfoRequest {
-    private BigInteger friendUserId;
+    private Long friendUserId;
     private String friendGroupName;
 }
