+++ src/main/java/com/qiao/picturedepot/service/FriendService.java
@@ -8,29 +8,28 @@ import com.qiao.picturedepot.pojo.dto.ApplyNewFriendRequest;
 import com.qiao.picturedepot.pojo.dto.FriendGroupDto;
 import com.qiao.picturedepot.pojo.dto.UpdateFriendInfoRequest;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface FriendService {
     /**
      * 获取按好友分组以及每组中的好友
      */
-    List<FriendGroupDto> getGroupedFriendList(BigInteger userId);
+    List<FriendGroupDto> getGroupedFriendList(Long userId);
 
     /**
      * 获取所有分组的名称列表
      */
-    List<String> getFriendGroupNames(BigInteger userId);
+    List<String> getFriendGroupNames(Long userId);
 
     /**
      * 检查两个用户是否存在好友关系
      */
-    boolean checkIsFriend(BigInteger userId1, BigInteger userId2);
+    boolean checkIsFriend(Long userId1, Long userId2);
 
     /**
      * 解除双向的好友关系
      */
-    void deleteFriend(BigInteger userId, BigInteger friendUserId);
+    void deleteFriend(Long userId, Long friendUserId);
 
     /**
      * 更新朋友分组信息（不可更新属主）
@@ -40,7 +39,7 @@ public interface FriendService {
     /**
      * 更新好友信息（分组）
      */
-    void updateFriendInfo(BigInteger userId, UpdateFriendInfoRequest updateFriendInfoRequest);
+    void updateFriendInfo(Long userId, UpdateFriendInfoRequest updateFriendInfoRequest);
 
     /**
      * 申请加为好友。发送申请消息。
@@ -50,10 +49,10 @@ public interface FriendService {
     /**
      * 接收好友申请。接收申请消息。
      */
-    void acceptNewFriend(BigInteger userId, AcceptNewFriendRequest acceptNewFriendRequest);
+    void acceptNewFriend(Long userId, AcceptNewFriendRequest acceptNewFriendRequest);
 
     /**
      * 拒绝好友申请。并向申请者发送拒绝提示。
      */
-    void rejectNewFriend(BigInteger userId, BigInteger systemMessageId);
+    void rejectNewFriend(Long userId, Long systemMessageId);
 }
