+++ src/main/java/com/qiao/picturedepot/service/UserService.java
@@ -2,11 +2,9 @@ package com.qiao.picturedepot.service;
 
 import com.qiao.picturedepot.pojo.dto.UserBaseInfoDto;
 
-import java.math.BigInteger;
-
 public interface UserService {
-    UserBaseInfoDto getUserBaseInfo(BigInteger userId);
-    BigInteger getUserIdByUsername(String username);
-    String getUsernameById(BigInteger userId);
+    UserBaseInfoDto getUserBaseInfo(Long userId);
+    Long getUserIdByUsername(String username);
+    String getUsernameById(Long userId);
     void registerUser(String username, String password);
 }
